requirements-dev.txt
antlr4-python3-runtime
codeclimate-test-reporter
coverage>4.0,<4.4
pep257
pre-commit
pre-commit
pylint<2.0
pytest
setuptools
